requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
prometheus-client==0.20.0
pillow==10.3.0
prometheus-fastapi-instrumentator==7.0.0
tensorflow==2.16.1
keras==3.3.3
numpy==1.26.0